Add unit tests for AI configuration resolution and provider functionality
